requirements.txt
# VHM24R - Интеллектуальная система сверки заказов, платежей и фискализации
# Зависимости для работы 24/7 на Railway с DigitalOcean

# Core Flask dependencies
Flask==2.3.2
gunicorn==21.2.0
python-dotenv==1.0.0

# Database support
psycopg2-binary==2.9.7
SQLAlchemy==2.0.21

# Data processing and Excel files
pandas==2.2.3
numpy==2.1.3
openpyxl==3.1.2
xlrd==2.0.1

# PDF processing
PyPDF2==3.0.1
pdfplumber==0.9.0

# Image processing and OCR
Pillow==10.0.0
pytesseract==0.3.10
opencv-python==4.8.0.76

# Additional file format support
python-docx==0.8.11

# Date and time handling
python-dateutil==2.8.2

# HTTP requests and API integration
requests==2.31.0

# DigitalOcean Spaces (S3-compatible storage)
boto3==1.28.62
botocore==1.31.62

# Telegram Bot API for notifications
python-telegram-bot==20.5

# Charting and visualization
matplotlib==3.7.2

# File handling utilities
chardet==5.2.0

# Security and utilities
Werkzeug==2.3.6
cryptography==41.0.4

# Cron job scheduling
APScheduler==3.10.4

# JSON handling
ujson==5.8.0

# Development and testing
pytest==7.4.0
pytest-flask==1.2.0

# Production monitoring
sentry-sdk[flask]==1.32.0

# Memory optimization
psutil==5.9.5
